More isolation: run_experiment.py, vs. design.py
run_experiment is entirely boilerplate universal code, to work in conjunction with data_management. Experiment specific code goes in design.py
