+++ 3_managing_design_results/experiment/run_experiment.py
@@ -1,78 +1,23 @@
-import sys
+"""Run a session of [this] experiment
+
+This module controls the basic experiment flow;
+finding incomplete trails/blocks, and iterating through them.
+
+The actual experimental procedure is defined in `design.py`
+
+Raises
+------
+SystemExit
+    if participant (/ experimenter) quits session before end
+"""
 
 import data_management
 import design
 import pandas as pd
 import text_displays
 from hrl import HRL
-from stimuli import sbc, whites
-
-# Define window parameters
-SHAPE = (1024, 1024)  # Desired shape of the drawing window
-CENTER = (SHAPE[0] // 2, SHAPE[1] // 2)  # Center of the drawing window
 
 
-def display_stim(ihrl, stim, intensity_target, intensity_left, intensity_right):
-    """Display stimulus with specified target, context intensities
-
-    Parameters
-    ----------
-    ihrl : hrl-object
-        hrl-interface object to use for display
-    stim : str
-        which stimulus to display, "sbc" or "whites"
-    intensity_target : float
-        intensity value for the targets
-    intensity_left : float
-        intensity value for the left context
-    intensity_right : float
-        intensity value for the right context
-    """
-    if stim == "sbc":
-        stimulus = sbc(intensity_target, intensity_left, intensity_right)
-    elif stim == "whites":
-        stimulus = whites(intensity_target, intensity_left, intensity_right)
-
-    # Convert the stimulus image(matrix) to an OpenGL texture
-    stim_texture = ihrl.graphics.newTexture(stimulus["img"])
-
-    # Determine position: we want the stimulus in the center of the frame
-    pos = (CENTER[1] - (stim_texture.wdth // 2), CENTER[0] - (stim_texture.hght // 2))
-
-    # Create a display: draw texture on the frame buffer
-    stim_texture.draw(pos=pos, sz=(stim_texture.wdth, stim_texture.hght))
-
-    # Display: flip the frame buffer
-    ihrl.graphics.flip(clr=True)  # also `clear` the frame buffer
-
-    return
-
-
-def respond(ihrl):
-    press, _ = ihrl.inputs.readButton(btns=("Left", "Right", "Escape", "Space"))
-
-    if press in ("Escape", "Space"):
-        # Raise SystemExit Exception
-        sys.exit("Participant terminated experiment.")
-    else:
-        return press
-
-
-def run_trial(ihrl, stim, intensity_target, intensity_left, intensity_right, **kwargs):
-    display_stim(ihrl, stim, intensity_target, intensity_left, intensity_right)
-    response = respond(ihrl)
-
-    if response == "Left":
-        result = intensity_left
-    elif response == "Right":
-        result = intensity_right
-
-    return {"response": response, "result": result}
-
-
-# ------------------------------------- #
-#           MAIN EXPERIMENT LOOP        #
-# ------------------------------------- #
 def run_block(hrl, block, block_id):
     print(f"Running block {block_id}")
     # Get start, end trial
@@ -93,7 +38,7 @@ def run_block(hrl, block, block_id):
 
         # run trial
         t1 = pd.Timestamp.now().strftime("%Y%m%d:%H%M%S.%f")
-        trial_results = run_trial(hrl, **trial)
+        trial_results = design.run_trial(hrl, **trial)
         trial.update(trial_results)
         t2 = pd.Timestamp.now().strftime("%Y%m%d:%H%M%S.%f")
 
@@ -125,7 +70,8 @@ def experiment_main(ihrl):
             print(f"Running session block {block_num+1}: {block_id}")
             block = run_block(ihrl, block=block, block_id=block_id)
 
-            text_displays.block_end(ihrl, block_num + 1, len(incomplete_blocks))
+            if block_num + 1 < len(incomplete_blocks):
+                text_displays.block_end(ihrl, block_num + 1, len(incomplete_blocks))
     except SystemExit as e:
         # Cleanup
         print("Exiting...")
@@ -144,8 +90,8 @@ if __name__ == "__main__":
         # graphics='datapixx',    # In the lab, we use the datapixx device driver
         inputs="keyboard",  # Use the keyboard as input device driver
         # inputs="responsepixx",  # In the lab, we use the responsepixx input device
-        hght=SHAPE[0],
-        wdth=SHAPE[1],
+        hght=design.SHAPE[0],
+        wdth=design.SHAPE[1],
         scrn=0,  # Which screen (monitor) to use
         fs=False,  # Fullscreen?
         bg=0.3,  # background intensity (black=0.0; white=1.0)
